+++ CHANGELOG.md
@@ -3,7 +3,10 @@
   - [#156](https://github.com/Datatamer/unify-client-python/issues/156) Fetch Dataset profile, even if out of date.
 
   **NEW FEATURES**
-  - [#65] (https://github.com/Datatamer/unify-client-python/issues/65) Fetches published clusters with data represented as a dataset.
+  - [#65](https://github.com/Datatamer/unify-client-python/issues/65) Fetches published clusters with data represented as a dataset.
+
+  **BUG FIXES**
+  - [#148](https://github.com/Datatamer/unify-client-python/issues/148) Fix null geo, id and absent id bug for geospatial datasets. 
 
 ## 0.6.0
   **BREAKING CHANGES**