+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - [#155](https://github.com/Datatamer/unify-client-python/issues/155) Adds read-only support for binning model.
   - [#165](https://github.com/Datatamer/unify-client-python/issues/165) Add `geo_attr` parameter to `Dataset.itergeofeatures()` and `Dataset.from_geo_features()`
   - [#113](https://github.com/Datatamer/unify-client-python/issues/113) Add support for uploading a binningModel
+  - [#168](https://github.com/Datatamer/unify-client-python/issues/168) Add support for project attributes
 
   **BUG FIXES**
   - [#148](https://github.com/Datatamer/unify-client-python/issues/148) Fix null geo, id and absent id bug for geospatial datasets.