+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - [#165](https://github.com/Datatamer/unify-client-python/issues/165) Add `geo_attr` parameter to `Dataset.itergeofeatures()` and `Dataset.from_geo_features()`
   - [#113](https://github.com/Datatamer/unify-client-python/issues/113) Add support for uploading a binningModel
   - [#168](https://github.com/Datatamer/unify-client-python/issues/168) Add support for project attributes
+  - [#171](https://github.com/Datatamer/unify-client-python/issues/171) Add support for creating and retrieving taxonomies
 
   **BUG FIXES**
   - [#148](https://github.com/Datatamer/unify-client-python/issues/148) Fix null geo, id and absent id bug for geospatial datasets.